+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.9.95-preview] - 2019-02-25
+- Add AssetStoreProvider example
+- Add HTTP asynchronous item loading by querying the asset store.
+- Add onEnable/onDisable API for Search provider.
+- Add Page Up and Down support to scroll the item list faster.
+- Add search provider examples
+- Add support for async results
+- Cycle too previous search query using ALT+UpArrow
+- Fix various Mac UX issues
+- New icons
+- Select first item by default en pressing Enter
+
 ## [0.9.9-preview.2] - 2019-02-21
 - Added drag and drop support. You can drag an item from the quick search tool to a drop target.
 - Open the item contextual menu by pressing the keyboard right arrow of the selected item.